gaussian_splatting: GPU harness Phase 2 — doctest-integrated runner
Replaces the Phase 1 spike with a doctest-integrated harness that runs
[RequiresGPU] cases against the offscreen RenderingDevice. The
HazardRepro test (test_output_compositor_composite_hazard.h) now
executes end-to-end on the godotgs Windows runner instead of skipping;
18/18 assertions pass, exit 0.

Harness (gs_gpu_test_runner.cpp, --gs-gpu-test):
- Main::test_setup() + RenderingContextDriverVulkan + RenderingDevice
  bootstrap (no RenderingServer needed — singleton self-registers in
  the RD constructor at rendering_device.cpp:8142, confirmed by Phase 1).
- doctest::Context argv pass-through mirroring tests/test_main.cpp:266-300,
  strips --gs-gpu-test and --gs-gpu-driver= before applying.
- Injects --test-case=*[RequiresGPU]* as the default filter when caller
  passes no --test-case= flag.
- GsGpuRidLeakListener registered globally; snapshots
  RD::get_memory_usage(MEMORY_TOTAL) per test case and emits an advisory
  print_line when delta > 1 MiB. No-op when get_singleton() is null so
  the listener stays quiet under the existing --test path.
- Exit-code contract: 0/1/2 from doctest, 64–66 for bootstrap failures,
  70 for test_setup failure, 71 if RD_ENABLED is not compiled.

Hazard test fixture additions:
- composite_with_destination=true alone falls through copy_to_render_target
  to copy_to_fb_rect (which needs CopyEffects::get_singleton() — not
  available headless). Setting depth_test_enabled=true + dummy 256x256
  D32_SFLOAT depth textures routes the test through the compute path
  where my Pass-2 scratch-copy fix actually lives.
- TEXTURE_USAGE_CAN_UPDATE_BIT added to source/destination format
  usage_bits so texture_update succeeds.
- Diagnostic prints converted from doctest's MESSAGE() to print_line()
  (MESSAGE captures const char* unsafely from temporary CharStrings).

visual_compare.h: %llu / %u → %d with int casts. Godot's vformat backs
onto String::sprintf, which doesn't support %llu or %u.

.gitignore: tests/visual_baselines/**/*.actual.png and
tests/visual_baselines/.provenance.json are now ignored so diff-capture
output and per-run provenance never land via casual git add.

First baseline: tests/visual_baselines/composite_hazard_256x256.png
captured on NVIDIA RTX 3090 / Vulkan via GS_VISUAL_BASELINE_MODE=update.
Compare-mode re-run passes against it; the gate is now live for any
local invocation. CI integration is Phase 3.

+++ modules/gaussian_splatting/tests/visual_compare.h
@@ -249,8 +249,8 @@ inline ComparisonResult compare_images(const Ref<Image> &p_a, const Ref<Image> &
 	}
 	result.match = (mismatched == 0);
 	result.diff_summary = vformat(
-			"mismatched=%u/%llu max_lsb=%d mean_lsb=%.3f psnr=%.2f dB",
-			mismatched, (unsigned long long)result.total_pixels,
+			"mismatched=%d/%d max_lsb=%d mean_lsb=%.3f psnr=%.2f dB",
+			int(mismatched), int(result.total_pixels),
 			max_lsb, result.mean_per_channel_diff_lsb, result.psnr_db);
 	return result;
 }
